refactoring(MovieQuizViewController): Move convert and question index functions to MovieQuizViewPresenter
